fixed type mismatch error

+++ src/main/java/com/pfe/ecredit/domain/DemandeHistorique.java
@@ -22,6 +22,6 @@ public class DemandeHistorique {
 	private Integer idTrt;
 	private Integer idPhase;
 	private LocalDate datePhase;
-	private Long idDemande;
+	private Integer idDemande;
 
 }
